+++ src/models/index.ts
@@ -1,3 +1,0 @@
-export * from './role';
-export * from './zone';
-export * from './permission';